updated IntersectionLayout index.js to include logic for generating random vehicles, including type and direction.
